% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CMAPfromDSEATM}
\alias{CMAPfromDSEATM}
\title{A dataframe including drugs and their related genes}
\format{
A dataframe with four columns from DSEATM
}
\usage{
CMAPfromDSEATM
}
\description{
A dataframe including drugs and their related genes
}
\keyword{datasets}
