% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{GSEA_dotplot}
\alias{GSEA_dotplot}
\title{Enrichment dotplot for positive or negative GSEA results}
\usage{
GSEA_dotplot(
  enrich.obj,
  type = c("pos", "neg"),
  show.term.num = 15,
  Selct.P = c("FDR", "P"),
  cutoff.P = 0.05,
  colors = c("blue", "white", "red"),
  size.range = c(3, 8),
  y.label.position = "right",
  title = NULL,
  legend.position = "right",
  ggtheme = theme_bw(base_rect_size = 1),
  ...
)
}
\arguments{
\item{enrich.obj}{A GSEA enrichment object from \code{clusterProfiler}.}

\item{type}{Specify whether you want to show positive or negative results.}

\item{show.term.num}{A number or a list of terms. If it is a number, the first n terms will be displayed. If it is a list of terms, the selected terms will be displayed.}

\item{Selct.P}{Selct.P P value (P) or adjust P value (FDR) were selected to define significant terms.}

\item{cutoff.P}{A cutoff value for \code{Select.P}.}

\item{colors}{A color vector for the dots.}

\item{size.range}{Two numeric variables, the first is minimal value and the first is maximal value.}

\item{y.label.position}{Y label position. right or left.}

\item{title}{Title of the plot.}

\item{legend.position}{Position of legend. 'none', 'right', 'left' or two numeric variables.}

\item{ggtheme}{ggtheme of plot.}

\item{...}{Other argument of \code{ggplot2::theme}.}
}
\description{
Plot enrichment dotplot for positive or negative GSEA results.
}
\examples{
\dontrun{
data(geneList, package="DOSE")
GSEA_res <- simple_GSEA(geneList, gene.type = "ENTREZID", enrich.type = 'GO')
GSEA_dotplot(res$simplyGO)
}



}
