% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA_intergated}
\alias{GSEA_intergated}
\title{Run GSEA analysis for all gene set}
\usage{
GSEA_intergated(
  genes,
  gene.type = "SYMBOL",
  organism = c("Human", "Mouse"),
  enrich.type = c("GO", "KEGG", "MKEGG", "WikiPathways", "Reactome", "MsigDB", "DO",
    "CGN", "DisGeNET", "CellMarker", "CMAP"),
  GO.ont = "BP",
  KEGG.use.internal.data = F,
  MsigDB.category = "H",
  pvalue.cutoff = 0.05,
  padjust.method = "BH",
  min.Geneset.Size = 10,
  max.Geneset.Size = 1000,
  CMAP.min.Geneset.Size = 3
)
}
\arguments{
\item{genes}{The same as in \code{simple_GSEA}.}

\item{gene.type}{The same as in \code{simple_GSEA}.}

\item{organism}{The same as in \code{simple_GSEA}.}

\item{enrich.type}{A vector of selected enrichment method. You can set one or more of GO, KEGG, MKEGG, WikiPathways, Reactome, MsigDB, DO, CGN, DisGeNET, CellMarker, and CMAP.}

\item{GO.ont}{The same as in \code{simple_GSEA}.}

\item{KEGG.use.internal.data}{The same as in \code{simple_GSEA}.}

\item{MsigDB.category}{The same as in \code{simple_GSEA}.}

\item{pvalue.cutoff}{The same as in \code{simple_GSEA}.}

\item{padjust.method}{The same as in \code{simple_GSEA}.}

\item{min.Geneset.Size}{The same as in \code{simple_GSEA}.}

\item{max.Geneset.Size}{The same as in \code{simple_GSEA}.}

\item{CMAP.min.Geneset.Size}{The same as in \code{simple_GSEA}.}
}
\description{
Perform integrated gene set enrichment analysis included GO, KEGG, WikiPathways, Reactome, MsigDB, Disease Ontoloty, Cancer Gene Network, DisGeNET, CellMarker, and CMAP.
}
\examples{
\dontrun{
  data(geneList, package="DOSE")

  # Set enrich.type using an enrichment analysis method mentioned above.
  fit <- GSEA_intergated(geneList, gene.type = 'ENTREZID', enrich.type = c("GO", "KEGG"))
}

}
