% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{GSEA_rankplot}
\alias{GSEA_rankplot}
\title{Visualize analyzing result of GSEA}
\usage{
GSEA_rankplot(
  GSEA.result,
  Pathway.ID,
  show.heatbar = T,
  show.rank = T,
  line.color = "#41A98E",
  rank.colors = c("#440154FF", "#482878FF", "#3E4A89FF", "#31688EFF", "#26828EFF",
    "#1F9E89FF", "#35B779FF", "#6DCD59FF", "#B4DE2CFF", "#FDE725FF"),
  heatbar.colors = c("#08519C", "#3182BD", "#6BAED6", "#BDD7E7", "#EFF3FF", "#FEE5D9",
    "#FCAE91", "#FB6A4A", "#DE2D26", "#A50F15"),
  add.x.ann = T,
  x.lab = "Gene ranks",
  line.y.lab = "Enrichment score",
  rank.y.lab = "logFC",
  statistic.position = c(0.5, 0.2),
  statistic.face = "italic",
  statistic.size = 3.5,
  ggtheme = theme_bw(base_rect_size = 1.5)
)
}
\arguments{
\item{GSEA.result}{GSEA results from \code{clusterProfiler::GSEA()} function.}

\item{Pathway.ID}{Pathway ID in the \code{ID} column of \code{GSEA.result} .}

\item{show.heatbar}{Whether show heat bar. Default TRUE.}

\item{show.rank}{Whether show Rank map. Default TRUE.}

\item{line.color}{Line color for running score.}

\item{rank.colors}{Color scheme of rank lines. A vector.}

\item{heatbar.colors}{Color scheme of heatbar. A vector.}

\item{add.x.ann}{Whether to add the title, text, and ticks of X axis.}

\item{x.lab}{x axis title.}

\item{line.y.lab}{Y label of running score plot.}

\item{rank.y.lab}{Y label of rank plot.}

\item{statistic.position}{Position of statistics in the running score plot.}

\item{statistic.face}{Font face of statistics.}

\item{statistic.size}{Font size of statistics.}

\item{ggtheme}{A theme object from ggplot2.}
}
\description{
Visualize analyzing result of GSEA.
}
\examples{
data(geneList, package="DOSE")
GSEA_res <- simple_GSEA(geneList, gene.type = "ENTREZID", enrich.type = 'KEGG')
GSEA_rankplot(GSEA_res, Pathway.ID = 'hsa04110')

}
