% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{GSEA_two_barplot}
\alias{GSEA_two_barplot}
\title{Enrichment barplot for two ORA enrichment objects}
\usage{
GSEA_two_barplot(
  enrich.obj,
  Selct.P = c("FDR", "P"),
  cutoff.P = 0.05,
  type.colors = c("#ED6355", "#3E94B5"),
  pos.top.pathway.num = 10,
  neg.top.pathway.num = 10,
  bar.width = 0.6,
  add.bar.border = T,
  x.limit.fold = 1.05,
  label.size = 3.5,
  legend.position = "bottom"
)
}
\arguments{
\item{enrich.obj}{An \code{gseaResult} object from clusterProfiler.}

\item{Selct.P}{P value (P) or adjust P value (FDR) were selected to define significant terms.}

\item{cutoff.P}{A cutoff value for \code{Select.P}.}

\item{type.colors}{Two colors for the positive and negative enrichment.}

\item{pos.top.pathway.num}{The number of top pathways in positive terms. Based on the significant test.}

\item{neg.top.pathway.num}{The number of top pathways in negative terms. Based on the significant test.}

\item{bar.width}{Width of bar in the plot.}

\item{add.bar.border}{Logical. Whether to add the black border of bars.}

\item{x.limit.fold}{Specify the fold of x limitation.}

\item{label.size}{Fontsize of label.}

\item{legend.position}{none, left, right, top, bottom; Or Two numeric variables indicated x and y positions, respectively.}
}
\description{
Plot enrichment barplot for two ORA enrichment objects.
}
\examples{
data(geneList, package="DOSE")

# Set enrich.type using an enrichment analysis method mentioned above.
fit <- simple_GSEA(geneList, enrich.type = "KEGG", gene.type = "ENTREZID")
GSEA_two_barplot(fit)

}
