% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ORA_dotplot}
\alias{ORA_dotplot}
\title{Plot Dotplot for enrihment result}
\usage{
ORA_dotplot(
  enrich.obj,
  x = "GeneRatio",
  color.by = "p.adjust",
  show.term.num = 15,
  label_format = 30,
  colors = c("blue", "white", "red"),
  color.title = color.by,
  size.by = "Count",
  size.range = c(3, 8),
  size.title = size.by,
  y.label.position = "right",
  title = NULL,
  legend.position = "right",
  ggtheme = theme_bw(base_rect_size = 1),
  ...
)
}
\arguments{
\item{enrich.obj}{An object from clusterProfiler.}

\item{x}{variable for x-axis, one of 'GeneRatio', 'pvalue', 'p.adjust', 'Count'.}

\item{color.by}{Variable that used to color enriched terms, one of 'GeneRatio', 'pvalue', 'p.adjust', 'Count'.}

\item{show.term.num}{A number or a list of terms. If it is a number, the first n terms will be displayed. If it is a list of terms, the selected terms will be displayed.}

\item{label_format}{a numeric value sets wrap length, alternatively a custom function to format axis labels. by default wraps names longer that 30 characters.}

\item{colors}{A color vector for the bars.}

\item{color.title}{Title of color annotation legend.}

\item{size.by}{Variable that used to size enriched terms, one of 'GeneRatio', 'pvalue', 'p.adjust', 'Count'}

\item{size.range}{Two numeric variables, the first is minimal value and the first is maximal value.}

\item{size.title}{Title of size annotation legend.}

\item{y.label.position}{Y label position. right or left.}

\item{title}{Title of the plot.}

\item{legend.position}{ostion of legend. 'none', 'right', 'left' or two numeric variables.}

\item{ggtheme}{ggtheme of plot.}

\item{...}{Other argument of \code{ggplot2::theme}.}
}
\description{
Plot Dotplot for enrihment result
}
\examples{
genes <- c("CANX", "HSPA1B", "KLRC2", "PSMC6", "RFXAP", "TAP1")
res <- simple_ORA(genes, enrich.type = "GO")
ORA_dotplot(res$simplyGO)

}
