% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA.R
\name{ORA_intergated}
\alias{ORA_intergated}
\title{Run over-representative analysis for all gene set}
\usage{
ORA_intergated(
  genes,
  background.genes = NULL,
  gene.type = "SYMBOL",
  organism = c("Human", "Mouse"),
  enrich.type = c("GO", "KEGG", "MKEGG", "WikiPathways", "Reactome", "MsigDB", "DO",
    "CGN", "DisGeNET", "CellMarker", "CMAP"),
  GO.ont = "BP",
  KEGG.use.internal.data = F,
  MsigDB.category = "H",
  pvalue.cutoff = 0.05,
  qvalue.cutoff = 0.05,
  padjust.method = "BH",
  min.Geneset.Size = 10,
  max.Geneset.Size = 1000,
  CMAP.min.Geneset.Size = 3
)
}
\arguments{
\item{genes}{The same as in \code{simple_ORA}.}

\item{background.genes}{The same as in \code{simple_ORA}.}

\item{gene.type}{The same as in \code{simple_ORA}.}

\item{organism}{The same as in \code{simple_ORA}.}

\item{enrich.type}{A vector of selected enrichment method. You can set one or more of GO, KEGG, MKEGG, WikiPathways, Reactome, MsigDB, DO, CGN, DisGeNET, CellMarker, and CMAP.}

\item{GO.ont}{The same as in \code{simple_ORA}.}

\item{KEGG.use.internal.data}{The same as in \code{simple_ORA}.}

\item{MsigDB.category}{The same as in \code{simple_ORA}.}

\item{pvalue.cutoff}{The same as in \code{simple_ORA}.}

\item{qvalue.cutoff}{The same as in \code{simple_ORA}.}

\item{padjust.method}{The same as in \code{simple_ORA}.}

\item{min.Geneset.Size}{The same as in \code{simple_ORA}.}

\item{max.Geneset.Size}{The same as in \code{simple_ORA}.}

\item{CMAP.min.Geneset.Size}{The same as in \code{simple_ORA}.}
}
\description{
Perform integrated over-representative enrichment analysis included GO, KEGG, WikiPathways, Reactome, MsigDB, Disease Ontoloty, Cancer Gene Network, DisGeNET, CellMarker, and CMAP.
}
\examples{
\dontrun{
 genes <- c("CANX", "HSPA1B", "KLRC2", "PSMC6", "RFXAP", "TAP1")
 res <- ORA_intergated(genes, enrich.type = c("GO", "KEGG"))
}

}
