% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ORA_two_barplot}
\alias{ORA_two_barplot}
\title{Enrichment barplot for two ORA enrichment objects}
\usage{
ORA_two_barplot(
  enrich.obj1,
  enrich.obj2,
  Selct.P = c("FDR", "P"),
  cutoff.P = 0.05,
  obj.types = c("Up", "Down"),
  obj.type.colors = c("#ED6355", "#3E94B5"),
  obj1.top.pathway.num = 10,
  obj2.top.pathway.num = 10,
  bar.width = 0.6,
  add.bar.border = T,
  x.limit.fold = 1.05,
  label.size = 3.5,
  legend.position = "bottom"
)
}
\arguments{
\item{enrich.obj1}{An \code{enrichResult} object from clusterProfiler.}

\item{enrich.obj2}{An \code{enrichResult} object from clusterProfiler.}

\item{Selct.P}{P value (P) or adjust P value (FDR) were selected to define significant terms.}

\item{cutoff.P}{A cutoff value for \code{Select.P}.}

\item{obj.types}{Two characters for defining the types of two objects.}

\item{obj.type.colors}{Two colors for the types of two objects.}

\item{obj1.top.pathway.num}{The number of top pathways in object 1. Based on the significant test.}

\item{obj2.top.pathway.num}{The number of top pathways in object 2. Based on the significant test.}

\item{bar.width}{Width of bar in the plot.}

\item{add.bar.border}{Logical. Whether to add the black border of bars.}

\item{x.limit.fold}{Specify the fold of x limitation.}

\item{label.size}{Fontsize of label.}

\item{legend.position}{none, left, right, top, bottom; Or Two numeric variables indicated x and y positions, respectively.}
}
\description{
Plot enrichment barplot for two ORA enrichment objects.
}
\examples{
genes <- c("CANX", "HSPA1B", "KLRC2", "PSMC6", "RFXAP", "TAP1")
res1 <- simple_ORA(genes, enrich.type = "KEGG")
res2 <- simple_ORA(genes, enrich.type = "MsigDB")
ORA_two_barplot(res1, res2, Selct.P = 'P', obj.types = c("KEGG", "MsigDB"))

}
