% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_gmt}
\alias{read_gmt}
\title{Read in gene set information from .gmt files}
\usage{
read_gmt(file)
}
\arguments{
\item{file}{The .gmt file to be read}
}
\value{
A list, where each index represents a separate gene set.
}
\description{
This function reads in and parses information from the MSigDB's .gmt files. Pathway information will be returned as a list of gene sets.
}
\details{
The .gmt format is a tab-delimited list of gene sets, where each line is a separate gene set. The first column must specify the name of the gene set, and the second column is used for a short description (which this function discards). For complete details on the .gmt format, refer to the Broad Institute's Data Format's page (url: http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats).
}
