% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{simple_KEGG_graph}
\alias{simple_KEGG_graph}
\title{KEGG pathway graph visualization}
\usage{
simple_KEGG_graph(
  gene.data,
  gene.type = "SYMBOL",
  pathway.id,
  species = "hsa",
  figure.suffix,
  save.dir = getwd()
)
}
\arguments{
\item{gene.data}{The same in \code{pathview::pathview()}.}

\item{gene.type}{The same in \code{pathview::pathview()}.}

\item{pathway.id}{The same in \code{pathview::pathview()}.}

\item{species}{The same in \code{pathview::pathview()}.}

\item{figure.suffix}{The same in \code{pathview::pathview()}.}

\item{save.dir}{The dir path to save the picture.}
}
\description{
Simple visualization of KEGG pathway based on pathview package.
}
