% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORA.R
\name{simple_ORA}
\alias{simple_ORA}
\title{Over-representative analysis}
\usage{
simple_ORA(
  genes,
  background.genes = NULL,
  gene.type = "SYMBOL",
  enrich.type = c("GO", "KEGG", "MKEGG", "WikiPathways", "Reactome", "MsigDB", "DO",
    "CGN", "DisGeNET", "CellMarker", "CMAP"),
  organism = c("Human", "Mouse"),
  GO.ont = c("BP", "CC", "MF", "ALL"),
  GO.simplify = T,
  KEGG.use.internal.data = F,
  MsigDB.category = "H",
  CMAP.min.Geneset.Size = 3,
  pvalue.cutoff = 0.05,
  qvalue.cutoff = 0.05,
  padjust.method = "BH",
  min.Geneset.Size = 10,
  max.Geneset.Size = 1000
)
}
\arguments{
\item{genes}{A vector of gene id.}

\item{background.genes}{Background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.}

\item{gene.type}{gene id type input gene.}

\item{enrich.type}{Select an enrichment method. One of GO, KEGG, MKEGG, WikiPathways, Reactome, MsigDB, DO, CGN, DisGeNET, CellMarker, and CMAP.}

\item{organism}{Specify species, currently support only Human and Mouse.}

\item{GO.ont}{GO parameter. One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{GO.simplify}{GO parameter. Whether to remove redundancy of enriched GO terms. If TRUE, will return a list with raw result and result after remove redundancy.}

\item{KEGG.use.internal.data}{KEGG parameter. Logical, use KEGG.db or latest online KEGG data.}

\item{MsigDB.category}{MsigDB parameter. MSigDB collection abbreviation, such as All, H, C1, C2, C3, C4, C5, C6, C7.}

\item{CMAP.min.Geneset.Size}{CMAP parameter. Minimal size of CMAP genes annotated for testing. Recommended use 3.}

\item{pvalue.cutoff}{pvalue cutoff on enrichment tests to report as significant.}

\item{qvalue.cutoff}{qvalue cutoff on enrichment tests to report as significant.}

\item{padjust.method}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{min.Geneset.Size}{Minimal size of genes annotated for testing. Not suitable for CMAP.}

\item{max.Geneset.Size}{Maximal size of genes annotated for testing.}
}
\description{
Perform over-representative analysis included GO, KEGG, WikiPathways, Reactome, MsigDB, Disease Ontoloty, Cancer Gene Network, DisGeNET, CellMarker, and CMAP.
}
\examples{
genes <- c("CANX", "HSPA1B", "KLRC2", "PSMC6", "RFXAP", "TAP1")
res <- simple_ORA(genes, enrich.type = "GO")

}
