% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{simple_ssgsea}
\alias{simple_ssgsea}
\title{Generate single-sample gene-set enrichment score}
\usage{
simple_ssgsea(exp, gene.list, method = c("ssgsea", "gsva", "zscore", "plage"))
}
\arguments{
\item{exp}{Numeric matrix containing the expression data or gene expression signatures, with samples in columns and genes in rows.}

\item{gene.list}{Gene sets provided either as a list object or as a GeneSetCollection object.}

\item{method}{Method to employ in the estimation of gene-set enrichment scores per sample.
By default this is set to gsva (Hänzelmann et al, 2013) and other options are ssgsea (Barbie et al, 2009), zscore (Lee et al, 2008) or plage (Tomfohr et al, 2005).
The latter two standardize first expression profiles into z-scores over the samples and,
in the case of zscore, it combines them together as their sum divided by the square-root of the size of the gene set,
while in the case of plage they are used to calculate the singular value decomposition (SVD) over the genes in the
gene set and use the coefficients of the first right-singular vector as pathway activity profile.}
}
\description{
Estimate gene-set enrichment score across all samples.
}
