% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbol.R
\name{update_symbol}
\alias{update_symbol}
\title{Update gene symbol by HGNC}
\usage{
update_symbol(genes, species = c("Human", "Mouse"), unmapGene_keep = F)
}
\arguments{
\item{genes}{a vector of gene symbol.}

\item{species}{A character vector of length 1, either "Human" (default) or "Mouse"}

\item{unmapGene_keep}{whether remove unmapped gene symbol.}
}
\description{
Update the gene symbol by HGNChelper
}
\examples{
human <- c("FN1", "TP53", "UNKNOWNGENE","7-Sep", "9/7", "1-Mar", "Oct4",
           "4-Oct", "OCT4-PG4", "C19ORF71", "C19orf71")
update_symbol(genes = human, unmapGene_keep = TRUE)

}
